refactor: markdown onLinkOpen
